AirSafe.com traffic spike analysis
Analysis of the pattern of traffic spikes on AirSafe.com, where spikes are defined as days when the number of sessions (as defined and measured by Google Analytics) on a particular day was higher than the average number of sessions in a 28-day comparison period by at least two standard deviations. Between 10 May 2006 and 14 November 2015, there were a total of 266 days with traffic spikes on AirSafe.com.